add:cambkos

+++ app/src/main/java/com/example/template2025/data/api/HomeResponse.kt
@@ -5,11 +5,16 @@ import com.google.gson.annotations.SerializedName
 data class HomeResponse(
     @SerializedName("Usuario") val usuario: Map<String, Any>?,
     @SerializedName("Dias") val dias: Map<String, Boolean>?,
+
     @SerializedName("Mision1") val mision1: Int?,
     @SerializedName("Mision2") val mision2: Int?,
     @SerializedName("Mision3") val mision3: Int?,
-    @SerializedName("Progreso") val progreso: Int?,
+
+    // Ahora como Double porque el backend manda decimal
+    @SerializedName("Progreso") val progreso: Double?,
+
     @SerializedName("Racha") val racha: Int?,
+
     @SerializedName("ProgresoModulo1") val progresoModulo1: Int?,
     @SerializedName("ProgresoModulo2") val progresoModulo2: Int?,
     @SerializedName("ProgresoModulo3") val progresoModulo3: Int?
